Changed layer copy-paste to use a JSON string on system clipboard. This allows for copying to/pasting from a text file which will make it easier for people to share single layers. Also made the most common JSONSerializerSettings be created in the Aurora.Utils.JSONUtils static class.
May want to consider moving the paste logic somewhere else, though currently the LayerList control is not actually aware whether it's data is for the normal layers or overlayers which makes it a tad awkward if this logic was to be moved to the ApplicationProfile class instead. Ideally there would be a LayerCollection (instead of an ObservableCollection<Layer>) which would then allow for adding specialized methods like this without needing to make the controls need to be aware of a bunch of other things.